feat(SimpleCard): preload data via property
