+++ test/components/SimpleCard/SimpleCard.spec.tsx
@@ -5,7 +5,7 @@ import { addHandlers } from "../../msw.setup"
 import { http, HttpResponse } from "msw"
 import { createElement } from "../../utils/jsx"
 import { createShopifyUrl } from "@/utils/createShopifyUrl"
-import type { ShopifyProduct } from "@/components/SimpleCard/types"
+import type { ShopifyProduct } from "@/shopify/types"
 
 describe("SimpleCard", () => {
   function addProductHandlers(responses: Record<string, { product?: ShopifyProduct; status?: number }>) {
